Split post 1425 oversized slices
