+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
@@ -2664,211 +2664,6 @@ private val VerticalSwipeThreshold = 72.dp
 private const val MotionDurationMs: Int = 200
 private val MotionEasing = CubicBezierEasing(0f, 0f, 0.2f, 1f)
 internal const val TMUX_SESSION_SCREEN_TAG = "tmux:session"
-internal const val TMUX_CONVERSATION_PANE_TAG = "tmux:conversation"
-
-// Issue #778: test tag on the conversation-loading placeholder shown when the
-// user taps Conversation before the transcript is ready. Issue #793 widened
-// this placeholder to a full load state machine (Loading / Failed / Empty).
-internal const val TMUX_CONVERSATION_DETECTING_TAG = "tmux:conversation:detecting"
-
-// Issue #793: test tags on the terminal conversation-load states (no infinite
-// spinner). Failed = clear error + retry; Empty = clear "no events" state.
-internal const val TMUX_CONVERSATION_LOAD_FAILED_TAG = "tmux:conversation:load-failed"
-internal const val TMUX_CONVERSATION_LOAD_RETRY_TAG = "tmux:conversation:load-retry"
-internal const val TMUX_CONVERSATION_LOAD_EMPTY_TAG = "tmux:conversation:load-empty"
-// Issue #1207: one-tap "Open in Terminal" action on the terminal-state
-// Conversation placeholder + the TUI-command notice — the only surface that can
-// drive a TUI-only slash-command picker (/model, /config …).
-internal const val TMUX_CONVERSATION_OPEN_TERMINAL_TAG = "tmux:conversation:open-terminal"
-internal const val TMUX_CONVERSATION_TUI_NOTICE_TAG = "tmux:conversation:tui-command-notice"
-internal const val TMUX_CONVERSATION_TUI_NOTICE_OPEN_TAG = "tmux:conversation:tui-command-open-terminal"
-// Issue #793: top-of-list progress row shown while older messages page in on
-// upward scroll (tail-first windowed load).
-internal const val TMUX_CONVERSATION_PAGING_OLDER_TAG = "tmux:conversation:paging-older"
-
-// Issue #423: actionable terminal-surface error state. Shown for a pane
-// whose local Termux surface failed to recover after a recovery storm
-// while SSH/tmux is still alive. The user taps the recreate control to
-// rebuild the surface without reconnecting SSH.
-internal const val TMUX_TERMINAL_SURFACE_ERROR_TAG = "tmux:terminal-surface-error"
-internal const val TMUX_TERMINAL_SURFACE_RECREATE_TAG = "tmux:terminal-surface-recreate"
-// Issue #459: the bespoke in-pane conversation composer (and its
-// composer-input / composer-send test tags) was removed — the Conversation
-// tab now shares the unified [PromptComposerSheet] at the screen level.
-internal const val TMUX_CONVERSATION_TOOL_ROW_TAG_PREFIX = "tmux:conversation:tool:"
-/** Issue #176: stable test tag prefix for a `SystemNote` row in the tmux conversation pane. */
-internal const val TMUX_CONVERSATION_SYSTEM_NOTE_ROW_TAG_PREFIX = "tmux:conversation:system-note:"
-/**
- * Issue #840 (slice 2): stable test tag prefix for the leading [DisclosureIcon]
- * on a `SystemNote` row's header — the shared expand/collapse affordance the row
- * previously lacked entirely.
- */
-internal const val TMUX_CONVERSATION_SYSTEM_NOTE_DISCLOSURE_TAG_PREFIX =
-    "tmux:conversation:system-note:disclosure:"
-/**
- * Issue #154/#786: stable test tags for the conversation navigation polish.
- * Issue #786 (hard-cut, D22): the "Search in conversation" field was DELETED —
- * the maintainer wants the whole screen given to the transcript — so its
- * `TMUX_CONVERSATION_SEARCH_TAG` is gone with it.
- * - [TMUX_CONVERSATION_JUMP_TO_LATEST_TAG] is on the "↓ Latest"
- *   affordance pinned to the bottom-end of the conversation list when
- *   the user has scrolled away from the tail. The button hides itself
- *   when the user is already at the bottom.
- */
-internal const val TMUX_CONVERSATION_JUMP_TO_LATEST_TAG =
-    "tmux:conversation:jump-to-latest"
-/**
- * Issue #154: test tag on the conversation feed's LazyColumn so the
- * navigation E2E test can drive scroll position directly via
- * `performScrollToIndex` (the indirection through a child text-row is
- * fragile because LazyColumn lazily disposes off-screen rows).
- */
-internal const val TMUX_CONVERSATION_LIST_TAG =
-    "tmux:conversation:list"
-/**
- * Issue #154: stable test tag for the brief pulse highlight that fades
- * over the Tabs row when the Conversation tab newly appears. The pulse
- * is a sibling overlay sitting on top of [TMUX_TABS_TAG]; tests can
- * `assertExists()` on this tag while the pulse animation runs and
- * `assertDoesNotExist()` after it completes.
- */
-internal const val TMUX_CONVERSATION_TAB_PULSE_TAG =
-    "tmux:tabs:conversation-pulse"
-
-/**
- * Issue #184 + #189: stable test tags for the IME-aware top chrome on
- * the tmux session screen.
- *
- * - [TMUX_FULL_BREADCRUMB_TAG] is on the IME-down consolidated 56dp
- *   toolbar (back chevron + status dot + `session › Window N` crumb +
- *   inline tab pill + kebab — all in one row, per issue #189). Pre-#189
- *   this tag was on the legacy 4-segment Breadcrumb that has been
- *   removed in favour of the consolidated row; the tag is preserved so
- *   the IME compression test in [TmuxSessionScreenImeChromeTest] keeps
- *   its "non-compressed chrome is displayed" assertion intact.
- * - [TMUX_COMPACT_BREADCRUMB_TAG] is on the slimmed-down session-name
- *   strip that replaces the consolidated row while the IME is up so the
- *   terminal viewport can claim the freed vertical space.
- * - [TMUX_TABS_TAG] is on the inline Terminal/Conversation toggle
- *   rendered inside [ConsolidatedTopChrome] (rendered only when an
- *   agent has been detected on the current window or a conversation is
- *   locked). Per #303 the toggle no longer consumes a dedicated row.
- * - [TMUX_CONSOLIDATED_SESSION_LABEL_TAG] tags the session crumb text
- *   inside the consolidated row so screenshot tests and future audits
- *   can find it without relying on translatable text.
- * - [TMUX_CONSOLIDATED_TAB_PILL_TAG_PREFIX] is the prefix for per-tab
- *   pill segments (index 0 = Terminal, index 1 = Conversation).
- * - [TMUX_TERMINAL_TAB_TAG] is a named, stable hook for the Terminal
- *   segment inside the toggle pill (equivalent to
- *   `TMUX_CONSOLIDATED_TAB_PILL_TAG_PREFIX + "0"`). Added for issue #216
- *   so connected E2E tests that previously asserted on the visible
- *   "Terminal" text can address the Terminal segment without depending
- *   on translatable strings. Note: the segment is only present when the
- *   pill itself is rendered (multi-tab case — i.e. an agent has been
- *   detected). For the universal "I am on the tmux session screen"
- *   sentinel, use [TMUX_SESSION_SCREEN_TAG] instead, since the pill is
- *   intentionally suppressed for shell-only / single-tab panes.
- */
-internal const val TMUX_FULL_BREADCRUMB_TAG = "tmux:breadcrumb:full"
-internal const val TMUX_COMPACT_BREADCRUMB_TAG = "tmux:breadcrumb:compact"
-internal const val TMUX_TABS_TAG = "tmux:tabs"
-internal const val TMUX_CONSOLIDATED_SESSION_LABEL_TAG = "tmux:chrome:session-label"
-internal const val TMUX_CONSOLIDATED_TAB_PILL_TAG_PREFIX = "tmux:chrome:tab-pill:"
-internal const val TMUX_TERMINAL_TAB_TAG = "tmux:chrome:tab-pill:terminal"
-internal const val TMUX_FULL_CHROME_BACK_BUTTON_TAG = "tmux:chrome:full:back"
-internal const val TMUX_FULL_CHROME_MORE_BUTTON_TAG = "tmux:chrome:full:more"
-internal const val TMUX_COMPACT_CHROME_BACK_BUTTON_TAG = "tmux:chrome:compact:back"
-internal const val TMUX_COMPACT_CHROME_MORE_BUTTON_TAG =
-    "tmux:chrome:compact:more"
-
-// Issue #462 regression sentinel: the old top-chrome command-palette button
-// must not render in either the full or compact tmux chrome. Kept so tests can
-// assert the previous affordance is gone from the header edge.
-internal const val TMUX_COMMAND_PALETTE_BUTTON_TAG = "tmux:chrome:command-palette"
-
-// Issue #463: the tappable project/folder crumb in the session header that
-// opens the in-session project-scoped session switcher dropdown, and the
-// dropdown's per-session rows.
-internal const val TMUX_PROJECT_SWITCHER_TAG = "tmux:chrome:project-switcher"
-internal const val TMUX_PROJECT_SWITCHER_MENU_TAG = "tmux:chrome:project-switcher:menu"
-internal const val TMUX_PROJECT_SWITCHER_ROW_TAG_PREFIX = "tmux:chrome:project-switcher:row:"
-
-/**
- * Issues #177 / #249: the "Reconnecting" / "Disconnected" breadcrumb pill
- * (see [ConnectionStatusPill]). Connected tests assert this appears while
- * the reattach handshake is in flight and disappears once the session is
- * live again.
- */
-internal const val TMUX_CONNECTION_STATUS_PILL_TAG = "tmux:chrome:connection-pill"
-
-
-/**
- * Issue #235: stable test tag for the kebab menu's Detach item. The
- * connected `TmuxDetachOnBackgroundE2eTest` does NOT click this tag —
- * it drives the lifecycle path via `UiDevice.pressHome` — but the
- * symmetric "manual detach + return to dashboard" assertion in the
- * same suite needs a deterministic handle on the menu item.
- */
-internal const val TMUX_DETACH_BUTTON_TAG = "tmux:session:detach-button"
-/**
- * Issue #445 (epic #432 slice A): stable test tag for the kebab menu's
- * "Port forwarding" item, used by the nav-route instrumentation to drive
- * kebab -> port-forward panel -> back-to-session.
- */
-internal const val TMUX_PORT_FORWARDING_BUTTON_TAG = "tmux:session:port-forwarding-button"
-/** Issue #592: stable test tag for the tmux kebab's global Settings item. */
-internal const val TMUX_SETTINGS_BUTTON_TAG = "tmux:session:settings-button"
-/**
- * Issue #892: stable test tag for the kebab's "Redraw" item. Tapping it forces a
- * full-viewport reseed of the active pane over the warm session (no reconnect /
- * detach / new lease) — the manual escape hatch from a black/partial-black terminal.
- */
-internal const val TMUX_REDRAW_BUTTON_TAG = "tmux:session:redraw-button"
-/**
- * Issue #993: stable test tag for the kebab's "Reconnect" item. Tapping it forces an
- * immediate reconnect of the CURRENT session in place (reusing the VM's single
- * [TmuxSessionViewModel.reconnect] / `TransportEffects.onManualReconnect` entrypoint) so
- * the user does NOT have to switch to another session and back to recover a dropped
- * session. On reconnect the #900 outbound queue auto-flushes the pending message. A
- * deliberate HALF-MEASURE escape hatch until auto-reconnect is bulletproof (#928).
- */
-internal const val TMUX_RECONNECT_BUTTON_TAG = "tmux:session:reconnect-button"
-/**
- * Issue #497: stable test tags for the kebab's "Open file…" item and the
- * path-entry dialog it opens, so instrumentation can drive
- * kebab -> enter path -> file viewer.
- */
-internal const val TMUX_OPEN_FILE_BUTTON_TAG = "tmux:session:open-file-button"
-internal const val TMUX_OPEN_FILE_DIALOG_FIELD_TAG = "tmux:session:open-file-field"
-internal const val TMUX_OPEN_FILE_DIALOG_CONFIRM_TAG = "tmux:session:open-file-confirm"
-internal const val TMUX_LIFECYCLE_DIALOG_CONFIRM_TAG = "tmux:session:lifecycle-dialog:confirm"
-internal const val TMUX_LIFECYCLE_DIALOG_CANCEL_TAG = "tmux:session:lifecycle-dialog:cancel"
-/**
- * Issue #528: stable test tag for the kebab's "Browse files…" item, used by
- * instrumentation to drive kebab -> file explorer.
- */
-internal const val TMUX_BROWSE_FILES_BUTTON_TAG = "tmux:session:browse-files-button"
-/**
- * Epic #821 Slice 1: stable test tag for the kebab's "Change kind" /
- * "What is this session?" item, used by instrumentation to open the
- * session-kind picker.
- */
-internal const val TMUX_CHANGE_KIND_BUTTON_TAG = "tmux:session:change-kind-button"
-/**
- * Issue #448 (epic #432 slice C): stable test tags for the new-port
- * detection overlay and its actions, so instrumentation can assert the
- * overlay appears and drive Forward / Dismiss.
- */
-internal const val TMUX_DETECTED_PORT_OVERLAY_TAG = "tmux:session:detected-port-overlay"
-internal const val TMUX_DETECTED_PORT_FORWARD_TAG = "tmux:session:detected-port-forward"
-internal const val TMUX_DETECTED_PORT_DISMISS_TAG = "tmux:session:detected-port-dismiss"
-internal const val TMUX_AGENT_QUICK_REPLY_ROW_TAG = "tmux:agent-quick-replies"
-internal const val TMUX_AGENT_QUICK_REPLY_CHIP_TAG_PREFIX = "tmux:agent-quick-reply:"
-
-internal const val TMUX_SESSION_PAGER_OVERLAY_TAG = "tmux:session-pager"
-internal const val TMUX_SESSION_PAGER_TAG = "tmux:session-pager-control"
-internal const val TMUX_SESSION_PAGER_PAGE_TAG_PREFIX = "tmux:session-pager-page:"
-internal const val TMUX_SESSION_PAGER_INDICATOR_TAG = "tmux:session-pager-indicator"
 
 private fun Modifier.verticalSwipeInput(
     thresholdPx: Float,
