+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionTags.kt
@@ -1,0 +1,70 @@
+package com.pocketshell.app.tmux
+
+internal const val TMUX_CONVERSATION_PANE_TAG = "tmux:conversation"
+
+internal const val TMUX_CONVERSATION_DETECTING_TAG = "tmux:conversation:detecting"
+
+internal const val TMUX_CONVERSATION_LOAD_FAILED_TAG = "tmux:conversation:load-failed"
+internal const val TMUX_CONVERSATION_LOAD_RETRY_TAG = "tmux:conversation:load-retry"
+internal const val TMUX_CONVERSATION_LOAD_EMPTY_TAG = "tmux:conversation:load-empty"
+
+internal const val TMUX_CONVERSATION_OPEN_TERMINAL_TAG = "tmux:conversation:open-terminal"
+internal const val TMUX_CONVERSATION_TUI_NOTICE_TAG = "tmux:conversation:tui-command-notice"
+internal const val TMUX_CONVERSATION_TUI_NOTICE_OPEN_TAG = "tmux:conversation:tui-command-open-terminal"
+internal const val TMUX_CONVERSATION_PAGING_OLDER_TAG = "tmux:conversation:paging-older"
+
+internal const val TMUX_TERMINAL_SURFACE_ERROR_TAG = "tmux:terminal-surface-error"
+internal const val TMUX_TERMINAL_SURFACE_RECREATE_TAG = "tmux:terminal-surface-recreate"
+internal const val TMUX_CONVERSATION_TOOL_ROW_TAG_PREFIX = "tmux:conversation:tool:"
+internal const val TMUX_CONVERSATION_SYSTEM_NOTE_ROW_TAG_PREFIX = "tmux:conversation:system-note:"
+internal const val TMUX_CONVERSATION_SYSTEM_NOTE_DISCLOSURE_TAG_PREFIX =
+    "tmux:conversation:system-note:disclosure:"
+internal const val TMUX_CONVERSATION_JUMP_TO_LATEST_TAG =
+    "tmux:conversation:jump-to-latest"
+internal const val TMUX_CONVERSATION_LIST_TAG =
+    "tmux:conversation:list"
+internal const val TMUX_CONVERSATION_TAB_PULSE_TAG =
+    "tmux:tabs:conversation-pulse"
+
+internal const val TMUX_FULL_BREADCRUMB_TAG = "tmux:breadcrumb:full"
+internal const val TMUX_COMPACT_BREADCRUMB_TAG = "tmux:breadcrumb:compact"
+internal const val TMUX_TABS_TAG = "tmux:tabs"
+internal const val TMUX_CONSOLIDATED_SESSION_LABEL_TAG = "tmux:chrome:session-label"
+internal const val TMUX_CONSOLIDATED_TAB_PILL_TAG_PREFIX = "tmux:chrome:tab-pill:"
+internal const val TMUX_TERMINAL_TAB_TAG = "tmux:chrome:tab-pill:terminal"
+internal const val TMUX_FULL_CHROME_BACK_BUTTON_TAG = "tmux:chrome:full:back"
+internal const val TMUX_FULL_CHROME_MORE_BUTTON_TAG = "tmux:chrome:full:more"
+internal const val TMUX_COMPACT_CHROME_BACK_BUTTON_TAG = "tmux:chrome:compact:back"
+internal const val TMUX_COMPACT_CHROME_MORE_BUTTON_TAG =
+    "tmux:chrome:compact:more"
+
+internal const val TMUX_COMMAND_PALETTE_BUTTON_TAG = "tmux:chrome:command-palette"
+
+internal const val TMUX_PROJECT_SWITCHER_TAG = "tmux:chrome:project-switcher"
+internal const val TMUX_PROJECT_SWITCHER_MENU_TAG = "tmux:chrome:project-switcher:menu"
+internal const val TMUX_PROJECT_SWITCHER_ROW_TAG_PREFIX = "tmux:chrome:project-switcher:row:"
+
+internal const val TMUX_CONNECTION_STATUS_PILL_TAG = "tmux:chrome:connection-pill"
+
+internal const val TMUX_DETACH_BUTTON_TAG = "tmux:session:detach-button"
+internal const val TMUX_PORT_FORWARDING_BUTTON_TAG = "tmux:session:port-forwarding-button"
+internal const val TMUX_SETTINGS_BUTTON_TAG = "tmux:session:settings-button"
+internal const val TMUX_REDRAW_BUTTON_TAG = "tmux:session:redraw-button"
+internal const val TMUX_RECONNECT_BUTTON_TAG = "tmux:session:reconnect-button"
+internal const val TMUX_OPEN_FILE_BUTTON_TAG = "tmux:session:open-file-button"
+internal const val TMUX_OPEN_FILE_DIALOG_FIELD_TAG = "tmux:session:open-file-field"
+internal const val TMUX_OPEN_FILE_DIALOG_CONFIRM_TAG = "tmux:session:open-file-confirm"
+internal const val TMUX_LIFECYCLE_DIALOG_CONFIRM_TAG = "tmux:session:lifecycle-dialog:confirm"
+internal const val TMUX_LIFECYCLE_DIALOG_CANCEL_TAG = "tmux:session:lifecycle-dialog:cancel"
+internal const val TMUX_BROWSE_FILES_BUTTON_TAG = "tmux:session:browse-files-button"
+internal const val TMUX_CHANGE_KIND_BUTTON_TAG = "tmux:session:change-kind-button"
+internal const val TMUX_DETECTED_PORT_OVERLAY_TAG = "tmux:session:detected-port-overlay"
+internal const val TMUX_DETECTED_PORT_FORWARD_TAG = "tmux:session:detected-port-forward"
+internal const val TMUX_DETECTED_PORT_DISMISS_TAG = "tmux:session:detected-port-dismiss"
+internal const val TMUX_AGENT_QUICK_REPLY_ROW_TAG = "tmux:agent-quick-replies"
+internal const val TMUX_AGENT_QUICK_REPLY_CHIP_TAG_PREFIX = "tmux:agent-quick-reply:"
+
+internal const val TMUX_SESSION_PAGER_OVERLAY_TAG = "tmux:session-pager"
+internal const val TMUX_SESSION_PAGER_TAG = "tmux:session-pager-control"
+internal const val TMUX_SESSION_PAGER_PAGE_TAG_PREFIX = "tmux:session-pager-page:"
+internal const val TMUX_SESSION_PAGER_INDICATOR_TAG = "tmux:session-pager-indicator"
